You will need to type in your password before the keyboard sends anything to the OS.
You will need to create a `secrets.c` and `secrets.h` file. The contents of which are shown in `sharedDefines.h`. This is so that you don't commit your password to github :). This also however prevents the firmware from building until those files are created.

Keyboard will also autolock after 5 minutes. It uses the same timer as the rgb timeout.

Add password entering animation if `#define SHOW_UNLOCK_ANIMATION`. But it should be commented out for increased security.

Keyboard can also be locked manually with Leader L.

Change rgb counters to 32 bit and add 5 and 10 minute timeouts.

Add CAPS WORD feature.

Change a few keys around in the keymap, and add a few media keys.